requirements.txt
torch
torchaudio
pyaudio
openai-whisper
